--- Frontend/front_gestor_turnos/src/(pages)/admin/page.tsx
@@ -1,9 +1,0 @@
-'use client'
-import React from 'react'
-import './pageAdmin.css'
-
-export default function Admin() {
-  return (
-    <div>page</div>
-  )
-}
