--- Frontend/front_gestor_turnos/src/(pages)/home/page.tsx
@@ -1,10 +1,0 @@
-'use clinet'
-import React from 'react';
-import './pageHome.css';
-
-
-export default function Home() {
-  return (
-    <div>Control de Ruteo - Page Home</div>
-  )
-}
