--- Frontend/front_gestor_turnos/src/(pages)/home/pageHome.css
@@ -1,5 +1,0 @@
-div {
-  background-color: aqua;
-  width: 100%;
-  height: 100px;
-}
