--- Frontend/front_gestor_turnos/src/(pages)/nadador/page.tsx
@@ -1,9 +1,0 @@
-'use client'
-import React from 'react'
-import './pageNadador.css'
-
-export default function Nadador() {
-  return (
-    <div>page</div>
-  )
-}
