--- Frontend/front_gestor_turnos/src/(pages)/profesor/profesor.tsx
@@ -1,9 +1,0 @@
-'use client'
-import React from 'react'
-import './pageProfesor.css'
-
-export default function Profesor() {
-  return (
-    <div>page</div>
-  )
-}
