+++ Frontend/front_gestor_turnos/src/app/(pages)/admin/page.tsx
@@ -1,0 +1,13 @@
+'use client'
+import React from 'react'
+import './pageAdmin.css'
+
+export default function Admin() {
+  return (
+    <>
+    <div className='div-pageAdmin'>Bienvenido ADMINISTRADOR</div>
+    {/* ../../home - Haciendo referencia a la carpeta a donde quiero volver */}
+    <a href='../../home'>Volver al inicio</a>
+    </>
+  )
+}
