+++ Frontend/front_gestor_turnos/src/app/(pages)/admin/pageAdmin.css
@@ -1,4 +1,4 @@
-div {
+.div-pageAdmin {
   background-color:burlywood;
   width: 100%;
   height: 100px;
