+++ Frontend/front_gestor_turnos/src/app/(pages)/home/page.tsx
@@ -1,0 +1,16 @@
+'use clinet'
+import React from 'react';
+import './pageHome.css';
+
+
+export default function Home() {
+  return (
+    <>
+    <div className='div-pageHome'>Control de Ruteo - Page Home</div>
+    <a href='../../admin'>Boton Admin</a>
+    <a href='../../profesor'>Boton Profesor</a>
+    <a href='../../nadador'>Boton Nadador</a>
+
+    </>
+  )
+}
