+++ Frontend/front_gestor_turnos/src/app/(pages)/home/pageHome.css
@@ -1,0 +1,5 @@
+.div-pageHome {
+  background-color: transparent;
+  width: 100%;
+  height: 100px;
+}
