+++ Frontend/front_gestor_turnos/src/app/(pages)/nadador/page.tsx
@@ -1,0 +1,14 @@
+'use client'
+import React from 'react'
+import './pageNadador.css'
+
+export default function Nadador() {
+  return (
+    <>
+    <div className='div-pageNadador'>Bienvenido Nadador</div>
+    {/* ../../home - Haciendo referencia a la carpeta a donde quiero volver */}
+    <a href='../../home'>Volver al inicio</a>
+    </>
+ 
+  )
+}
