+++ Frontend/front_gestor_turnos/src/app/(pages)/nadador/pageNadador.css
@@ -1,4 +1,4 @@
-div {
+.div-pageNadador {
   background-color:blue;
   width: 100%;
   height: 100px;
