+++ Frontend/front_gestor_turnos/src/app/(pages)/profesor/page.tsx
@@ -1,0 +1,13 @@
+'use client'
+import React from 'react'
+import './pageProfesor.css'
+
+export default function Profesor() {
+  return (
+    <>
+    <div className='div-pageProfesor'>Bienvenido Profesor</div>
+    {/* ../../home - Haciendo referencia a la carpeta a donde quiero volver */}
+    <a href='../../home'>Volver al inicio</a>
+    </>
+  )
+}
