+++ Frontend/front_gestor_turnos/src/app/(pages)/profesor/pageProfesor.css
@@ -1,4 +1,4 @@
-div {
+.div-pageProfesor {
   background-color:blueviolet;
   width: 100%;
   height: 100px;
