+++ Frontend/front_gestor_turnos/src/app/componentes/footer/Footer.css
@@ -1,5 +1,5 @@
-div {
+.div-Footer {
   background-color: cadetblue;
   width: 100%;
-  height: 100px;
+  height: 20px;
 }
