+++ Frontend/front_gestor_turnos/src/app/componentes/footer/Footer.tsx
@@ -5,6 +5,6 @@ import './Footer.css'
 
 export const Footer = () => {
   return (
-    <div>Header</div>
+    <div className='div-Footer'>Footer</div>
   )
 }
