+++ Frontend/front_gestor_turnos/src/app/componentes/header/Header.css
@@ -1,5 +1,5 @@
-div {
+.div-Header{
   background-color:chartreuse;
   width: 100%;
-  height: 100px;
+  height: 20px;
 }
