+++ Frontend/front_gestor_turnos/src/app/componentes/header/Header.tsx
@@ -5,6 +5,6 @@ import './Header.css'
 
 export const Header = () => {
   return (
-    <div>Header</div>
+    <div className='div-Header'>Header</div>
   )
 }
