+++ Frontend/front_gestor_turnos/src/app/layout.tsx
@@ -1,9 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import { Footer } from "../componentes/footer/Footer";
-import { Header } from "../componentes/header/Header";
-import "./globals.css";
-
+import { Footer } from "../app/componentes/footer/Footer";
+import { Header } from "../app/componentes/header/Header";
 
 
 const inter = Inter({ subsets: ["latin"] });
@@ -15,13 +13,13 @@ export const metadata: Metadata = {
 
 export default function RootLayout({
   children,
-}: Readonly<{
+}: {
   children: React.ReactNode;
-}>) {
+}) {
   return (
     <>
       <html lang="en">
-      <body className={inter.className}><Header />{children}<Footer /></body>
+        <body className={inter.className}><Header/>{children}<Footer/></body>
       </html>
     </>
   );
